feat(ai-agent): let agent write back the active changeset
Add a `fromActiveChangeset` flag to the `proposeWriteback` agent tool so
the agent can open a pull request from a project's staged changeset when
the user asks (e.g. "create a PR from my changesets"). The writeback
instructions are built deterministically server-side from the
changeset's structured changes.

Removes the now-redundant ChangesetWritebackController and
ChangesetWritebackService, which exposed the same behaviour via a hidden
API endpoint that nothing called.

Closes: PROD-8127
