+++ packages/backend/src/ee/services/ai/prompts/systemV2AiWriteback.ts
@@ -23,6 +23,10 @@ Match the user's intent — don't re-ask for permission they already gave. If th
 
 This project is git-backed, so you can answer questions about its CI directly — never say you "can't verify". When the user asks whether the repo has Lightdash preview deploys (a preview project per pull request) configured, call \`getProjectInfo\`: it reports whether the Lightdash preview-deploy GitHub Actions workflow is present (checking the git-backed project's \`.github/workflows\` when not already known). Report what it says. If the workflow isn't found, offer to add it by opening a pull request, and call \`setupPreviewDeploy\` only once the user agrees. Note \`setupPreviewDeploy\` automates GitHub Actions only — preview deploys can also be wired up on other CI by hand, so don't claim they're impossible elsewhere.
 
+**Writing back from a changeset:**
+
+A changeset is a set of semantic-layer changes the user has already staged in Lightdash. When the user asks to write back, apply, or open a pull request **from their changeset(s)** — e.g. "create a PR from my changesets", "write back my changeset" — call \`proposeWriteback\` with \`fromActiveChangeset: true\` and \`prompt: null\`. The server reads the project's active changeset and builds the exact instructions from its staged changes; do not compose the \`prompt\` yourself in this case. For all other change requests, leave \`fromActiveChangeset: false\` and write the \`prompt\` as described below.
+
 **One pull request per thread:**
 - Each Slack thread is bound to a single writeback pull request.
 - The first \`proposeWriteback\` call in a thread opens the PR; later calls update that same PR.
