+++ packages/common/src/ee/AiAgent/schemas/tools/toolProposeWritebackArgs.ts
@@ -11,15 +11,20 @@ export const TOOL_PROPOSE_WRITEBACK_DESCRIPTION = [
 export const toolProposeWritebackArgsSchema = z.object({
     prompt: z
         .string()
-        .min(1)
+        .nullable()
         .describe(
-            'A focused, self-contained natural-language instruction for the writeback agent describing exactly which files in the dbt project to change and how. The writeback agent does not see this conversation, so include every detail it needs (model name, file path hints, the literal change to make). Do not include preamble or pleasantries.',
+            'A focused, self-contained natural-language instruction for the writeback agent describing exactly which files in the dbt project to change and how. The writeback agent does not see this conversation, so include every detail it needs (model name, file path hints, the literal change to make). Do not include preamble or pleasantries. Pass null only when fromActiveChangeset is true, in which case this is ignored.',
         ),
     prUrl: z
         .string()
         .nullable()
         .describe(
             "If the user pasted a link to an existing GitHub pull request they want to UPDATE instead of opening a new one, put the full PR URL here (e.g. 'https://github.com/owner/repo/pull/123'). The PR must belong to this project's own dbt repository. Only set this when the user explicitly references an existing PR to edit; otherwise pass null to open a new pull request.",
+        ),
+    fromActiveChangeset: z
+        .boolean()
+        .describe(
+            'Set to true when the user asks to write back, apply, or open a pull request FROM their changeset(s) — e.g. "create a PR from my changesets". The server then reads the project\'s active changeset and builds the writeback instructions deterministically from its structured changes, ignoring `prompt` (pass null). Leave false for ordinary free-text writeback requests where you compose `prompt` yourself.',
         ),
 });
